rsem v1.1.13, speed up EM by only updating model parameters for first 10 iterations, skip reads with its (or at least one of its mates', for paired-end reads) length < 25bp

+++ buildReadIndex.cpp
@@ -53,7 +53,7 @@ void buildIndex(char* readF, int gap, bool hasQ) {
 		}
 		++nReads;
 
-		if (verbose && nReads % 1000000 == 0) { printf("FIN %lld\n", nReads); }
+		if (verbose && nReads % 1000000 == 0) { printf("FIN %lld\n", (long long)nReads); }
 	} while (success);
 
 	fout.seekp(startPos);
